feat/CRS-gui-update-student-result

+++ Function/GetAll/GetAllScores/GetAllScores.cpp
@@ -15,10 +15,10 @@ void readScore(std::ifstream &fin, Score &score) {
     getline(fin, midtermMark);
     getline(fin, finalMark);
     getline(fin, totalMark);
-    score.otherMark = stoi(otherMark);
-    score.midtermMark = stoi(midtermMark);
-    score.finalMark = stoi(finalMark);
-    score.totalMark = stoi(totalMark);
+    score.otherMark = stod(otherMark);
+    score.midtermMark = stod(midtermMark);
+    score.finalMark = stod(finalMark);
+    score.totalMark = stod(totalMark);
 }
 
 Node<Score> *getAllScores() { return getAll("Data/Score.txt", &readScore); }
